requirements.txt
sphinx
sphinx-autobuild
invoke
streamlit
pycryptodome